Module HetHashtbl

Most of this code is adapated from the OCaml Standard library's implementation of Hashtbl. It is just adapted for polymorphic types. I.E. a 'b t hashtabl binding 'a key to ('a,'b) value (note that 'a is quantified existantially here).

Copyright 1996 Institut National de Recherche en Informatique et en Automatique. distributed under the terms of the GNU Lesser General Public License version 2.1, with the special exception on linking described in the license

module type HETEROGENEOUS_HASHED_TYPE = sig ... end

module type HETEROGENEOUS_SEEDED_HASHED_TYPE = sig ... end

val is_randomized : unit -> bool

module type S = sig ... end

module MakeSeeded (Key : HETEROGENEOUS_SEEDED_HASHED_TYPE) (Value : PatriciaTree.HETEROGENEOUS_VALUE) : S with type 'key key = 'key Key.t and type ('key, 'a) value = ('key, 'a) Value.t

module Make (Key : HETEROGENEOUS_HASHED_TYPE) (Value : PatriciaTree.HETEROGENEOUS_VALUE) : S with type 'key key = 'key Key.t and type ('key, 'a) value = ('key, 'a) Value.t